requirements.txt
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dwf==0.1.0
fonttools==4.51.0
h5py==3.11.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
Pylon==0.4.4
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
pyqtgraph==0.13.4
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scipy==1.13.0
six==1.16.0
